move create_oauth_jwt to utils.py
